+++ src/proxy/win-sso.facade.factory.ts
@@ -10,8 +10,8 @@ export class WinSsoFacadeFactory implements IWinSsoFacadeFactory {
     securityPackage: string,
     targetHost: string | undefined,
     peerCert: PeerCertificate | undefined,
-    delegate: boolean
+    flags: number | undefined
   ): IWinSsoFacade {
-    return new WinSso(securityPackage, targetHost, peerCert, delegate);
+    return new WinSso(securityPackage, targetHost, peerCert, flags);
   }
 }