+++ src/proxy/win-sso.facade.ts
@@ -8,8 +8,8 @@ export class WinSsoFacade implements IWinSsoFacade {
   private _winSso: WinSso;
 
   constructor(securityPackage: string, targetHost: string | undefined,
-    peerCert: PeerCertificate | undefined, delegate: boolean) {
-    this._winSso = new WinSso(securityPackage, targetHost, peerCert, delegate);
+    peerCert: PeerCertificate | undefined, flags: number | undefined) {
+    this._winSso = new WinSso(securityPackage, targetHost, peerCert, flags);
   }
 
   createAuthRequestHeader(): string {